+++ lib/bloc/keyboard_shift/keyboard_shift_event.dart
@@ -1,0 +1,20 @@
+part of 'keyboard_shift_bloc.dart';
+
+abstract class KeyboardShiftEvent extends Equatable {
+  const KeyboardShiftEvent();
+
+  @override
+  List<Object> get props => [];
+}
+
+class KeyboardShiftUpperCaseEvent extends KeyboardShiftEvent {
+  const KeyboardShiftUpperCaseEvent();
+}
+
+class KeyboardShiftLowerCaseEvent extends KeyboardShiftEvent {
+  const KeyboardShiftLowerCaseEvent();
+}
+
+class KeyboardShiftSymbolsEvent extends KeyboardShiftEvent {
+  const KeyboardShiftSymbolsEvent();
+}