+++ test/onscreen_keyboard_test.dart
@@ -1,9 +1,7 @@
 import 'package:flutter_test/flutter_test.dart';
 
-import 'package:onscreen_keyboard/onscreen_keyboard.dart';
+
 
 void main() {
-  test('adds one to input values', () {
-    final calculator = OnscreenKeyboard();
-  });
+  test('adds one to input values', () {});
 }