+++ CHANGELOG.md
@@ -1,3 +1,9 @@
-## [0.0.1] - Wed May 13, 2020
+## [1.0.1] - Wed Sept 16, 2020
 
-* This is the initial release
+* Add uppercase, lowercase and special characters
+
+
+## [1.0.0] - Wed May 13, 2020
+
+* Add autofocus to space bar button and format code 
+